Updates since 3.0
[UPD] CMake option BTK_EXTRA_COMPILER_WARNINGS works also with the Clang compiler.

Clang warnings: Unused variables.

Update the code of Open3DMotion to the commit c74cde4528 (part 1/2).

Update the code of Open3DMotion to the commit c74cde4528 (part 2/2).

Clang complains about signed/unsigned comparison.

Some unit tests failed under Linux 32-bit (compiled from a 64-bit machine) due to some round off errors. Adding the flag -ffloat-store solved most of the problems. The tolerance of one unit test had to be changed (from 1e-14 to 5e-14). It is not known if this fix is necessary if the code is built with a true 32-bit processor.

Update the code of Open3DMotion to the commit 21033cdb1a (part 1/3).

Update the code of Open3DMotion to the commit 21033cdb1a (part 2/3).

Update the code of Open3DMotion to the commit 21033cdb1a (part 3/3).

Copyright years updated.

New methods to extract signed/unsigned 64-bit integer from a binary file stream.

New predefined analog gain +/-1.65V.

Cleanup related to the predefined analog gains. Some predefined gains were removed from the C3D writer as they are not defined in the specification of the file format.

New method btk::Analog::SetGainFromValue.

Cleanup in the Delsys EMG file reader to use the method btk::Analog::SetGainFromValue.

[ENH] New reader for the Delsys file format HPF.

BTKCore set to the development version number “0.4dev”.

[FIX] Issue #76 (Crash of the acquisition reader when a MDF file has no marker).

The team behind Travis may update the build environment under MacOS X as CMake is now installed by default (see build 47 -- https://travis-ci.org/Biomechanical-ToolKit/BTKCore/builds/18427607).

Compiler error with GCC when the CMake option BTK_EXTRA_COMPILER_WARNINGS is activated.

[FIX] In case of a merging (and not a concatenation), the values for the points and the analog channels are not always correctly written (wrong number of frames to write).

[FIX] Bug in the OT XLS reader when the number of column's data is greater than the number of labels. Fix also an issue when the spatiotemporal parameters are empty as well as the events.

Because it seems GCC 4.7 does not like the Eigen::Matrix::isApprox method and because this method is not exactly what we want (we want to compare the value of each coefficient and not the Hilbert-Schmidt norm), the use of this function was replaced by some adapted code.

Following the commit 26a6af3, the unit tests are adapted to not use the method Eigen::Matrix::isApprox.

[FIX] Missing a test to know if the given corner's coordinate is equal than the stored one.

[FIX] Issue 77: "btkGetPoints has an error as pointsInfo.residuals returns an empty structure"

[UPD] FindMatlab updated to detected Matlab r2013(a|b) versions.

[UPD] FindMatlab updated to detected Matlab r2014a version.

[UPD] The analog offset is now stored as a double floating-point instead of an integer. This gives the possibility to store "fake" offset for digital sensors. The C3D file format was adapted in consequence to write the parameter ANALOG:OFFSET with the format Real if necessary.

The Python wrapping of the class btk::C3DFileIO was not adapted in the commit f5890c1.

Files reordered alphabetically.

[FIX] Wrong type use in the reinterpret_cast function.

Format HPF not listed in the table of supported acquisition file formats.

Some code in the IO module refactorized to be able later to create other binary stream/device (buffer, serial port, etc).

Macro btkChooseNativeBinaryFileStream renamed _btk_choose_native_binary_file_stream as it is only for internal usages.

Second part of the code factorization to read/write data into selected binary byte order stream outside of the class btk::BinaryFileStream.

Exception messages detailed to distinguish when it happens during the reading or the writing of a C3D file.

The code related the method btk::mmfilebuf::is_eob is removed as the logic behind this buffer cannot never reach this state.

Reuse of the variable 'n'.

It seems not necessary to remap the file buffer if the location of the written bytes goes afterwards exactly to the end of the buffer (position + number == buffer size).

Members related to a (random access) binary stream moved into the class btk::BinaryStream.

[UPD] TRB file reader improved to extract unnamed marker. This modification also resolves a crash related to the storage of unknown markers.

[FIX] The Matlab function btkCropAcquistion throws an error when the acquisition has no analog channel to crop (Luca Modenese).

This checks seems to be necessary because if the frames are calculated elsewhere in Matlab numFrames and startAt are generally integer numbers in "double" format. The original btkCropAcquisition function gives an error if index are not in integer format.

a fix to my wrong fix :-)

a fix to my wrong fix :-)

Added check on input variables

CMake seems no more installed by default in MacOS X environment. Plus, the gcc compiler entry was removed as it is not installed by default.

Commit 40ab4d7 breaks a unit test. The code was shifted in consequence.

[FIX] Wrong data extraction in the header of the Kistler DAT file format. The proposed modifications might be wrong as the comprehension of the bytes in header are not all known.

Upgrade eigen to the latest version. The previous version does not build with MSVC.